Change variables from uint32_t to uint8_t and tidy

+++ Rotary/rotary.cpp
@@ -187,7 +187,7 @@ void update_button(uint32_t *button_history){   // Called by Systick every 1ms
 }
 
 uint8_t is_button_up(uint32_t *button_history){
-    // Returns true of button up
+    // Returns true if button up
     return(*button_history == 0x00000000); 
 }
 
@@ -198,20 +198,22 @@ uint8_t is_button_down(uint32_t *button_history){
 
 uint8_t is_button_pressed(uint32_t *button_history){
     // Returns true if pressed
-    uint32_t pressed = 0;
+    uint8_t pressed = 0;
     if ((*button_history & BUTTON_MASK) == 0b00000000000000000000000011111111){
         pressed = 1;
-        *button_history = 0xFFFFFFFF; // Set Histtory to prevent retriggering from same press
+        // Set History to prevent retriggering from same press
+        *button_history = 0xFFFFFFFF; 
     }
     return pressed;
 }
 
 uint8_t is_button_released(uint32_t *button_history){
     // Returns true if released
-    uint32_t released = 0;
+    uint8_t released = 0;
     if ((*button_history & BUTTON_MASK) == 0b11111111000000000000000000000000){
         released = 1;
-        *button_history = 0x00000000; // Clear Histtory to prevent retriggering from same release
+        // Clear Histtory to prevent retriggering from same release
+        *button_history = 0x00000000; 
     }
     return released;
 }
